Fix various other small bugs

+++ core/src/main/java/com/crschnick/pdx_unlimiter/core/info/stellaris/StellarisSavegameInfo.java
@@ -26,7 +26,7 @@ public class StellarisSavegameInfo extends SavegameInfo<StellarisTag> {
 
             i.binary = false;
 
-            int seed = n.getNodeForKey("random_seed").getInteger();
+            long seed = n.getNodeForKey("random_seed").getLong();
             byte[] b = new byte[20];
             new Random(seed).nextBytes(b);
             i.campaignHeuristic = UUID.nameUUIDFromBytes(b);
